+++ packages/configs/eslint-ts-react/index.js
@@ -14,44 +14,48 @@ module.exports = {
         {
             files: ['**/*.ts', '**/*.tsx'],
             parser: '@typescript-eslint/parser',
-            plugins: ['@typescript-eslint', 'log-filenames', 'prettier'],
+            plugins: [
+                '@typescript-eslint',
+                'react',
+                'react-hooks',
+                'jest',
+                'log-filenames',
+                'prettier'
+            ],
             extends: [
                 'eslint:recommended',
                 'plugin:@typescript-eslint/recommended',
-                // 'plugin:@typescript-eslint/recommended-requiring-type-checking', // fails to initialise with vs code
-                // 'plugin:jest/recommended',
-                // 'plugin:react/recommended', //< covered in next
-                // 'plugin:react-hooks/recommended', //< covered in next
-                // 'plugin:@next/next/recommended', //< covered in next
-                'next', // https://github.com/vercel/next.js/blob/canary/packages/eslint-config-next/package.json
-                'next/core-web-vitals',
+                'plugin:@typescript-eslint/recommended-requiring-type-checking',
+                'plugin:jest/recommended',
+                'plugin:react/recommended',
+                'plugin:react-hooks/recommended',
                 'prettier'
             ],
             rules: {
                 '@typescript-eslint/no-empty-interface': 'off',
                 '@typescript-eslint/no-explicit-any': 'off',
                 '@typescript-eslint/no-namespace': 'off',
+                //'@typescript-eslint/no-non-null-assertion': 'off',
                 '@typescript-eslint/no-var-requires': 'off',
-                '@typescript-eslint/no-unused-vars': 'warn',
-                'react/react-in-jsx-scope': 'off',
+                '@typescript-eslint/ban-types': [
+                    'error',
+                    {
+                        extendDefaults: true,
+                        types: {
+                            '{}': false
+                        }
+                    }
+                ],
                 'react-internal/no-production-logging': 'off',
                 'react/display-name': 'off',
+                'react/react-in-jsx-scope': 'off',
                 '@next/next/no-img-element': 'off'
-            },
-            env: {
-                browser: true,
-                node: true,
-                jest: true
             },
             parserOptions: {
                 ecmaVersion: 'latest',
-                sourceType: 'module',
-                project: ['./apps/*/', './packages/*/*/']
+                project: ['./packages/*/*/tsconfig.json']
             },
             settings: {
-                next: {
-                    rootDir: ['apps/*/', 'packages/*/*/']
-                },
                 react: {
                     version: 'detect' // Tells eslint-plugin-react to automatically detect the version of React to use
                 }