+++ apps/tetris-classic/CHANGELOG.md
@@ -1,5 +1,16 @@
 # tetris-classic
 
+## 0.3.1
+
+### Patch Changes
+
+- Update docs and versions
+- Updated dependencies
+  - @muravjev/configs-ts@0.1.1
+  - @muravjev/features-fallback@0.1.0
+  - @muravjev/utils-deferred@0.1.0
+  - @muravjev/utils-redux-subslice@0.1.1
+
 ## 0.3.0
 
 ### Minor Changes