+++ packages/utils/redux-subslice/CHANGELOG.md
@@ -1,7 +1,13 @@
 # @muravjev/utils-redux-subslice
 
+## 0.1.1
+
+### Patch Changes
+
+- Update docs and versions
+
 ## 0.1.0
 
 ### Minor Changes
 
-- Extract utils packages
+- Create initial version